requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
blinker==1.8.2
boto3==1.34.145
botocore==1.34.145
Brlapi==0.8.5
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
click==8.1.7
cloud-init==24.1.3
cohere==5.6.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.2.1
cryptography==41.0.7
cupshelpers==1.0
cvlearn==0.1.3
cycler==0.12.1
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
faiss-cpu==1.8.0.post1
fastavro==1.9.5
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
greenlet==3.0.3
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
idna==3.6
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
mysql-connector-python==9.0.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
openai==1.35.14
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pexpect==4.9.0
pillow==10.4.0
protobuf==4.25.3
ptyprocess==0.7.0
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu1
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.2
scipy==1.14.0
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
SQLAlchemy==2.0.31
systemd-python==235
termcolor==2.4.0
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.32.0.20240712
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
Werkzeug==3.0.3
wheel==0.42.0
xdg==5
xkit==0.0.0
